+++ src/datatable/datatables-timer.ts
@@ -14,12 +14,9 @@ export default class DatatablesTimer {
       console.log('loaded all tabs in', Math.round(duration) + ' ms')
       pageContentLoaded.set(true)
 
-      let route = ''
-      if (window.location.hash) {
-        route = window.location.hash.split('#/')[1].split('?')[0]
-      } else {
-        route = window.location.pathname.substring(1)
-      }
+      const route = window.location.hash
+        ? window.location.hash.split('#/')[1].split('?')[0]
+        : window.location.pathname.substring(1)
       pageLoadedRoute.set(route.replace(/\//g, '___'))
     }
   }